webstorage/storage-removeItem-change-order.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WebStorage Test: Storage - removeItem change order</title>
<link rel="author" title="Intel" href="http://www.intel.com" >
<link rel="help" href="https://w3c.github.io/webstorage/#the-storage-interface" >
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  ["localStorage", "sessionStorage"].forEach(function(name) {
    test(function () {
      var storage = window[name];
      storage.clear();

      storage.setItem("key1", "value1");
      storage.setItem("key2", "value2");
      var key = storage.key(0);
      assert_equals(key, "key1");
      assert_equals(storage.getItem(key), "value1");

      storage.removeItem(key); 
      assert_equals(storage.key(0), "key2"); 

    }, name + ": Check if Storage.removeItem can change the order of keys.");
  });
</script>
